requirements.txt
fastapi
uvicorn
pydantic
neo4j
gitpython
python-dotenv
openai
qdrant-client[fastembed]
networkx
pytest